+++ src/brod_consumer.erl
@@ -22,12 +22,14 @@
         , start_link/4
         , start_link/5
         , stop/1
+        , stop_maybe_kill/2
         , subscribe/3
         , unsubscribe/2
         ]).
 
 %% Debug API
 -export([ debug/2
+        , get_connection/1
         ]).
 
 %% gen_server callbacks
@@ -67,7 +69,7 @@
 
 -type pending_acks() :: #pending_acks{}.
 
--record(state, { client_pid          :: pid()
+-record(state, { bootstrap           :: pid() | brod:bootstrap()
                , connection          :: ?undef | pid()
                , topic               :: binary()
                , partition           :: integer()
@@ -112,10 +114,10 @@
 
 %%%_* APIs =====================================================================
 %% @equiv start_link(ClientPid, Topic, Partition, Config, [])
--spec start_link(pid(), topic(), partition(), config()) ->
+-spec start_link(pid() | brod:bootstrap(), topic(), partition(), config()) ->
         {ok, pid()} | {error, any()}.
-start_link(ClientPid, Topic, Partition, Config) ->
-  start_link(ClientPid, Topic, Partition, Config, []).
+start_link(Bootstrap, Topic, Partition, Config) ->
+  start_link(Bootstrap, Topic, Partition, Config, []).
 
 %% @doc Start (link) a partition consumer.
 %% Possible configs:
@@ -157,14 +159,27 @@ start_link(ClientPid, Topic, Partition, Config) ->
 %%     then let it dynamically adjust to a number around
 %%     PrefetchCount * AverageSize
 %% @end
--spec start_link(pid(), topic(), partition(), config(), [any()]) ->
+-spec start_link(pid() | brod:bootstrap(),
+                 topic(), partition(), config(), [any()]) ->
                     {ok, pid()} | {error, any()}.
-start_link(ClientPid, Topic, Partition, Config, Debug) ->
-  Args = {ClientPid, Topic, Partition, Config},
+start_link(Bootstrap, Topic, Partition, Config, Debug) ->
+  Args = {Bootstrap, Topic, Partition, Config},
   gen_server:start_link(?MODULE, Args, [{debug, Debug}]).
 
 -spec stop(pid()) -> ok | {error, any()}.
 stop(Pid) -> safe_gen_call(Pid, stop, infinity).
+
+-spec stop_maybe_kill(pid(), timeout()) -> ok.
+stop_maybe_kill(Pid, Timeout) ->
+  try
+    gen_server:call(Pid, stop, Timeout)
+  catch
+    exit : {noproc, _} ->
+      ok;
+    exit : {timeout, _} ->
+      exit(Pid, kill),
+      ok
+  end.
 
 %% @doc Subscribe or resubscribe on messages from a partition.
 %% Caller may pass in a set of options which is an extention of consumer config
@@ -204,9 +219,13 @@ debug(Pid, print) ->
 debug(Pid, File) when is_list(File) ->
   do_debug(Pid, {log_to_file, File}).
 
+%% @doc Get connection pid. Test/debug only.
+get_connection(Pid) ->
+  gen_server:call(Pid, get_connection).
+
 %%%_* gen_server callbacks =====================================================
 
-init({ClientPid, Topic, Partition, Config}) ->
+init({Bootstrap, Topic, Partition, Config}) ->
   erlang:process_flag(trap_exit, true),
   Cfg = fun(Name, Default) ->
           proplists:get_value(Name, Config, Default)
@@ -219,8 +238,14 @@ init({ClientPid, Topic, Partition, Config}) ->
   PrefetchBytes = erlang:max(Cfg(prefetch_bytes, ?DEFAULT_PREFETCH_BYTES), 0),
   BeginOffset = Cfg(begin_offset, ?DEFAULT_BEGIN_OFFSET),
   OffsetResetPolicy = Cfg(offset_reset_policy, ?DEFAULT_OFFSET_RESET_POLICY),
-  ok = brod_client:register_consumer(ClientPid, Topic, Partition),
-  {ok, #state{ client_pid          = ClientPid
+  %% If bootstrap is a client pid, register self to the client
+  case is_shared_conn(Bootstrap) of
+    true ->
+      ok = brod_client:register_consumer(Bootstrap, Topic, Partition);
+    false ->
+      ok
+  end,
+  {ok, #state{ bootstrap           = Bootstrap
              , topic               = Topic
              , partition           = Partition
              , begin_offset        = BeginOffset
@@ -267,20 +292,27 @@ handle_info({'DOWN', _MonitorRef, process, Pid, _Reason},
                                      }),
   {noreply, NewState};
 handle_info({'DOWN', _MonitorRef, process, Pid, _Reason},
-            #state{connection = Pid} = State0) ->
-  State = State0#state{connection = ?undef, connection_mref = ?undef},
-  ok = maybe_send_init_connection(State),
-  {noreply, State};
+            #state{connection = Pid} = State) ->
+  %% monitored connection managed by brod_client
+  {noreply, handle_conn_down(State)};
+handle_info({'EXIT', Pid, _Reason}, #state{connection = Pid} = State) ->
+  %% standalone connection spawn-linked to self()
+  {noreply, handle_conn_down(State)};
 handle_info(Info, State) ->
   error_logger:warning_msg("~p ~p got unexpected info: ~p",
                           [?MODULE, self(), Info]),
   {noreply, State}.
 
+handle_call(get_connection, _From, #state{connection = C} = State) ->
+  {reply, C, State};
 handle_call({subscribe, Pid, Options}, _From,
             #state{subscriber = Subscriber} = State0) ->
   case (not brod_utils:is_pid_alive(Subscriber)) %% old subscriber died
     orelse Subscriber =:= Pid of                 %% re-subscribe
     true ->
+      %% Ensure connection is established before replying this call
+      %% because we may need the connection
+      %% to resolve begin offset (latest/earliest)
       case maybe_init_connection(State0) of
         {ok, State} ->
           handle_subscribe_call(Pid, Options, State);
@@ -318,24 +350,35 @@ handle_cast(Cast, State) ->
                           [?MODULE, self(), Cast]),
   {noreply, State}.
 
-terminate(Reason, #state{client_pid = ClientPid
+terminate(Reason, #state{ bootstrap = Bootstrap
                         , topic = Topic
                         , partition = Partition
+                        , connection = Connection
                         }) ->
-  case brod_utils:is_normal_reason(Reason) of
-    true ->
-      brod_client:deregister_consumer(ClientPid, Topic, Partition);
-    false ->
-      ok
+  IsShared = is_shared_conn(Bootstrap),
+  IsNormal = brod_utils:is_normal_reason(Reason),
+  %% deregister consumer if it's shared connection and normal shutdown
+  IsShared andalso IsNormal andalso
+    brod_client:deregister_consumer(Bootstrap, Topic, Partition),
+  %% close connection if it's working standalone
+  case not IsShared andalso is_pid(Connection) of
+    true -> kpro:close_connection(Connection);
+    false -> ok
   end,
-  error_logger:warning_msg("~p ~p terminating, reason:\n~p",
-                           [?MODULE, self(), Reason]),
+  %% write a log if it's not shared connection
+  IsNormal orelse error_logger:error_msg("Consumer ~s-~w terminate reason: ~p",
+                                         [Topic, Partition, Reason]),
   ok.
 
 code_change(_OldVsn, State, _Extra) ->
   {ok, State}.
 
 %%%_* Internal Functions =======================================================
+
+handle_conn_down(State0) ->
+  State = State0#state{connection = ?undef, connection_mref = ?undef},
+  ok = maybe_send_init_connection(State),
+  State.
 
 do_debug(Pid, Debug) ->
   {ok, _} = gen:call(Pid, system, {debug, Debug}, infinity),
@@ -728,15 +771,19 @@ safe_gen_call(Server, Call, Timeout) ->
 %% Init payload connection regardless of subscriber state.
 -spec maybe_init_connection(state()) ->
         {ok, state()} | {{error, any()}, state()}.
-maybe_init_connection(#state{ client_pid = ClientPid
-                        , topic      = Topic
-                        , partition  = Partition
-                        , connection = ?undef
-                        } = State0) ->
+maybe_init_connection(
+  #state{ bootstrap  = Bootstrap
+        , topic      = Topic
+        , partition  = Partition
+        , connection = ?undef
+        } = State0) ->
   %% Lookup, or maybe (re-)establish a connection to partition leader
-  case brod_client:get_leader_connection(ClientPid, Topic, Partition) of
+  case connect_leader(Bootstrap, Topic, Partition) of
     {ok, Connection} ->
-      Mref = erlang:monitor(process, Connection),
+      Mref = case is_shared_conn(Bootstrap) of
+               true -> erlang:monitor(process, Connection);
+               false -> ?undef %% linked
+             end,
       %% Switching to a new connection
       %% the response for last_req_ref will be lost forever
       State = State0#state{ last_req_ref = ?undef
@@ -745,11 +792,19 @@ maybe_init_connection(#state{ client_pid = ClientPid
                           },
       {ok, State};
     {error, Reason} ->
-      {{error, Reason}, State0}
+      {{error, {connect_leader, Reason}}, State0}
   end;
 maybe_init_connection(State) ->
   {ok, State}.
 
+
+connect_leader(ClientPid, Topic, Partition) when is_pid(ClientPid) ->
+  brod_client:get_leader_connection(ClientPid, Topic, Partition);
+connect_leader(Endpoints, Topic, Partition) when is_list(Endpoints) ->
+  connect_leader({Endpoints, []}, Topic, Partition);
+connect_leader({Endpoints, ConnCfg}, Topic, Partition) ->
+  %% connection pid is linked to self()
+  kpro:connect_partition_leader(Endpoints, ConnCfg, Topic, Partition).
 
 %% Send a ?INIT_CONNECTION delayed loopback message to re-init.
 -spec maybe_send_init_connection(state()) -> ok.
@@ -759,6 +814,9 @@ maybe_send_init_connection(#state{subscriber = Subscriber}) ->
   brod_utils:is_pid_alive(Subscriber) andalso
     erlang:send_after(Timeout, self(), ?INIT_CONNECTION),
   ok.
+
+%% In case 'bootstrap' is a client pid, connection is shared with other workers.
+is_shared_conn(Bootstrap) -> is_pid(Bootstrap).
 
 %%%_* Tests ====================================================================